refactored docker ops into a separate module
